proposed refactoring: move InteractiveRenderer to a module
